Add ability to supersede a previous ADR record (#2)
* Start work on 1.0.1 with the ability to supersede an ADR.
* Complete Supersede functionality and refactor some small bits

+++ Adr/AdrTool.cs
@@ -10,35 +10,54 @@ namespace Adr;
 
 public sealed class AdrTool
 {
-    private readonly string _rootFolder = string.Empty;
-    private readonly string _docsFolder = string.Empty;
-    private string? _indexFile;
+    private readonly SortedList<int, AdrEntry> _adrEntries = new();
+    private readonly string _gitRootPath = string.Empty;
+    private readonly string _adrFolder = string.Empty;
+    private readonly string? _indexFile;
 
     public AdrTool(string givenPath)
     {
+        if (string.IsNullOrEmpty(givenPath) || givenPath == ".")
+        {
+            givenPath = Environment.CurrentDirectory;
+        }
+
         if (!Path.Exists(givenPath))
         {
             Cout.Fail("Error: The given path '{Path}' does not exist", givenPath);
             return;
         }
 
-        if (string.IsNullOrEmpty(givenPath) || givenPath == ".")
-        {
-            givenPath = Environment.CurrentDirectory;
-        }
-
-        _rootFolder = GetGitRootFolder(givenPath);
-        if (string.IsNullOrEmpty(_rootFolder))
+        _gitRootPath = GetGitRootFolder(givenPath);
+        if (string.IsNullOrEmpty(_gitRootPath))
         {
             Cout.Fail("Error: The provided Path '{ProvidedPath}' is not within a git repository", givenPath);
             Aborted = true;
             return;
         }
 
-        _docsFolder = GetDocsFolder(givenPath);
-        if (string.IsNullOrEmpty(_docsFolder))
+        _adrFolder = GetAdrFolder(givenPath);
+
+        if (string.IsNullOrEmpty(_adrFolder))
         {
-            _docsFolder = AskForAndCreateAdrFolder();
+            _adrFolder = AskForAndCreateAdrFolder();
+
+            var newEntry = new AdrEntry
+            {
+                Number = 1,
+                Title = "Record Architecture Decision Records",
+                FilePath = "0001-record-architecture-decision-records.md",
+                SupersededBy = 0
+            };
+
+            DocumentCreator.Create(_adrFolder, newEntry, DocumentCreator.DecisionToUseAdrAsMarkdown);
+
+            _adrEntries.Add(1, newEntry);
+            _indexFile = IndexManipulator.CreateIndexFromAdrEntries(_adrEntries, _adrFolder);
+        }
+        else
+        {
+            _indexFile = Path.Combine(_adrFolder, "0000-index.md");
         }
     }
 
@@ -47,48 +66,86 @@ public sealed class AdrTool
     public void Run()
     {
         DisplayInfo();
-        var filesInFolder = Directory.GetFiles(_docsFolder, "*.md");
-        _indexFile = Array.Find(filesInFolder, f => f.Contains("0000-index"));
 
-        if (string.IsNullOrEmpty(_indexFile))
+        var entries = IndexManipulator.ReadAdrEntries(_adrFolder);
+        foreach (var entry in entries)
         {
-            Cout.Fail("Ubable to locate ADR index in '{Path}'", _docsFolder);
-            return;
+            if (!_adrEntries.ContainsKey(entry.Number))
+            {
+                _adrEntries.Add(entry.Number, entry);
+            }
         }
 
-        var entries = IndexManipulator.GetEntriesFromIndexFile(_indexFile);
+        var tableMenu = new[]
+        {
+            new LiveKeyAction<AdrEntry>('a', "Add new", _ => AppendNew()),
+            new LiveKeyAction<AdrEntry>('r', "Rename selected", entry => Rename(entry)),
+            new LiveKeyAction<AdrEntry>('i', "Recreate index from folder", _ => IndexManipulator.RecreateIndex(_adrFolder)),
+            new LiveKeyAction<AdrEntry>('o', "Open ADR folder in VS Code", _ => VSCode.OpenFolder(_adrFolder)),
+            new LiveKeyAction<AdrEntry>('s', "Supersede with new ADR entry", e => Supersede(e))
+        };
 
-        if (entries.Any())
+        if (_adrEntries.Any())
         {
             new LiveDataTable<AdrEntry>()
-                .WithHeader($"There are {entries.Count} Architecture decision(s) made so far in this solution\nADR path: [yellow]{_docsFolder}[/]\n")
+                .WithHeader($"There are {entries.Count()} Architecture decision(s) made so far in this solution\nADR path: [yellow]{_adrFolder}[/]\n")
                 .WithDataSource(entries)
                 .WithColumns("Id", "Title")
                 .WithDataPicker(e => new List<string> { e.Number.ToString(CultureInfo.InvariantCulture), e.Title })
                 .WithEnterInstruction("open ADR #{0} in VS Code.\nUse arrow up/down to select.\n", p => p.Number.ToString(CultureInfo.InvariantCulture))
-                .WithMultipleActions(new[]
-                {
-                    new LiveKeyAction<AdrEntry>('a', "Add new", _ => AppendNew(entries)),
-                    new LiveKeyAction<AdrEntry>('r', "Rename selected", entry => Rename(entry, entries)),
-                    new LiveKeyAction<AdrEntry>('i', "Recreate index from folder", _ => IndexManipulator.RecreateIndex(_docsFolder)),
-                    new LiveKeyAction<AdrEntry>('o', "Open ADR folder in VS Code", _ => VSCode.OpenFolder(_docsFolder))
-                })
-                .WithSelectionAction(entry => VSCode.OpenFile(Path.Combine(_docsFolder, entry.Url)))
-
+                .WithMultipleActions(tableMenu)
+                .WithSelectionAction(entry => VSCode.OpenFile(Path.Combine(_adrFolder, entry.FilePath)))
                 .Start();
         }
     }
 
+    private void Supersede(AdrEntry existingEntry)
+    {
+        Cout.Info("Supersede Entry titled: {existing}", existingEntry.Title);
+        var confirmd = AnsiConsole.Confirm($"Create a new Entry that supersedes #{existingEntry.Number}?", defaultValue: false);
+        if (!confirmd)
+        {
+            Cout.Warn("Supersede '{title}' aborted", existingEntry.Title);
+            return;
+        }
+
+        var title = AnsiConsole.Ask<string>("[yellow]New title:[/] ", existingEntry.Title);
+        var newEntry = CreateNewAdrFromTitle(title);
+        existingEntry.SupersededBy = newEntry.Number;
+
+        _adrEntries.Add(newEntry.Number, newEntry);
+
+        var created = DocumentCreator.Create(_adrFolder, newEntry);
+
+        IndexManipulator.CreateIndexFromAdrEntries(_adrEntries, _adrFolder);
+
+        VSCode.OpenFile(created);
+    }
+
+    private AdrEntry CreateNewAdrFromTitle(string title)
+    {
+        var nextNumber = _adrEntries.Max(e => e.Key) + 1;
+        var fileNamePart = title.Replace(" ", "-", StringComparison.OrdinalIgnoreCase).ToLower(CultureInfo.InvariantCulture).Trim();
+
+        return new AdrEntry
+        {
+            Number = nextNumber,
+            Title = title,
+            FilePath = $"./{nextNumber:0000}-{fileNamePart}.md",
+            SupersededBy = 0
+        };
+    }
+
     private string AskForAndCreateAdrFolder()
     {
-        if (!AnsiConsole.Confirm("Did not find ADR document folder structure. Should I create it?", defaultValue: false))
+        if (!AnsiConsole.Confirm("Did not find ADR document folder structure.\nDo you want to create it now?", defaultValue: false))
         {
             Cout.Info("Ok, aborting.");
             Aborted = true;
             return string.Empty;
         }
 
-        var adrFolder = Path.Combine(_rootFolder, "Docs");
+        var adrFolder = Path.Combine(_gitRootPath, "Docs");
         if (!Directory.Exists(adrFolder))
         {
             Directory.CreateDirectory(adrFolder);
@@ -100,23 +157,22 @@ public sealed class AdrTool
         {
             Directory.CreateDirectory(adrFolder);
             Cout.Success("Created folder: {Folder}", adrFolder);
-
-            IndexManipulator.CreateInitialFolderContent(adrFolder);
         }
 
         return adrFolder;
     }
 
-    private void Rename(AdrEntry entry, List<AdrEntry> entries)
+    private void Rename(AdrEntry entry)
     {
         var originalTitle = entry.Title;
-        var originalUrl = entry.Url;
-        var originalFilePath = Path.Combine(_docsFolder, originalUrl);
+        var originalUrl = entry.FilePath;
+        var originalFilePath = Path.Combine(_adrFolder, originalUrl);
 
-        var newTitle = AnsiConsole.Ask<string>("New title: ");
+        Cout.Info("Rename entry: {EntryTitle}\n", entry.Title);
+        var newTitle = AnsiConsole.Ask<string>("[yellow]New title:[/] ", entry.Title);
         entry.Title = newTitle;
-        entry.Url = $"./{entry.Number:0000}-{newTitle.Replace(" ", "-").ToLower(CultureInfo.InvariantCulture)}.md";
-        var destinationPath = Path.Combine(_docsFolder, entry.Url);
+        entry.FilePath = $"./{entry.Number:0000}-{newTitle.Replace(" ", "-").ToLower(CultureInfo.InvariantCulture)}.md";
+        var destinationPath = Path.Combine(_adrFolder, entry.FilePath);
 
         var content = File.ReadAllText(originalFilePath);
         Cout.Info("Opened {file}", originalFilePath);
@@ -127,50 +183,60 @@ public sealed class AdrTool
         File.WriteAllText(destinationPath, content);
         Cout.Info("Saved as '{file}'", destinationPath);
 
-        File.Delete(originalFilePath);
-        Cout.Info("Deleted '{file}'", originalFilePath);
+        if (!destinationPath.Equals(originalFilePath, StringComparison.OrdinalIgnoreCase))
+        {
+            File.Delete(originalFilePath);
+            Cout.Info("Deleted '{file}'", originalFilePath);
+        }
 
-        IndexManipulator.Rewrite(entries, _indexFile!);
+        IndexManipulator.Rewrite(_adrEntries, _indexFile!);
     }
 
-    private void AppendNew(List<AdrEntry> allEntries)
+    private void AppendNew()
     {
-        var nextNumber = allEntries.Max(e => e.Number) + 1;
-        var newName = AnsiConsole.Ask<string>("Enter a [yellow]title[/] for the new ADR: ");
-        var newFileName = $"{nextNumber:0000}-{newName.Replace(" ", "-").ToLower(CultureInfo.InvariantCulture)}.md";
+        var nextNumber = _adrEntries.Max(e => e.Value.Number) + 1;
+        Cout.Info("Create New");
+        var newName = AnsiConsole.Ask<string>("Enter a descriptive [yellow]Title[/] for your ADR: ");
+        var newFileName = $"./{nextNumber:0000}-{newName.Replace(" ", "-").ToLower(CultureInfo.InvariantCulture)}.md";
         var newEntry = new AdrEntry
         {
             Number = nextNumber,
             Title = newName,
-            Url = newFileName
+            FilePath = newFileName,
+            SupersededBy = 0
         };
 
-        Cout.Success("New entry #{Number} titled '{Title}' will be created as {Path}", newEntry.Number, newEntry.Title, newEntry.Url);
+        Cout.Success("Entry #{Number}: '{Title}' will be created as {Path}", newEntry.Number, newEntry.Title, newEntry.FilePath);
 
         if (AnsiConsole.Confirm("Do you want to Proceed?", defaultValue: false))
         {
-            var created = DocumentCreator.Create(_docsFolder, newEntry);
+            var created = DocumentCreator.Create(_adrFolder, newEntry);
             if (string.IsNullOrEmpty(created))
             {
                 Cout.Fail("Somehow, ths failed. I'm at a loss here!");
                 return;
             }
 
-            allEntries.Add(newEntry);
-            IndexManipulator.Rewrite(allEntries, _indexFile!);
+            _adrEntries.Add(newEntry.Number, newEntry);
+            IndexManipulator.Rewrite(_adrEntries, _indexFile!);
             VSCode.OpenFile(created);
+        }
+        else
+        {
+            Cout.Warn("No document created. Program finished");
         }
     }
 
     private static void DisplayInfo()
     {
+        Cout.Cls();
         Cout.Hr("ADR Tool v1.0.0 - 2023 - digitaldias");
         Cout.Info(" ");
     }
 
-    private string GetDocsFolder(string? givenPath)
+    private string GetAdrFolder(string? givenPath)
     {
-        while (!string.IsNullOrEmpty(givenPath) && BothPathsShareRoot(givenPath, _rootFolder))
+        while (!string.IsNullOrEmpty(givenPath) && BothPathsShareRoot(givenPath, _gitRootPath))
         {
             var sampler = Path.Combine(givenPath, "docs", "adr");
             if (Directory.Exists(sampler))
